chore: move config into `/internal`
